moved PriceFactory from ReadModel to Framework to be accessible for FE API
